etc/msynth/requirements.txt
pandas
pathlib
faker
pyyaml
croniter
pyarrow
faker_airtravel